src/requirements.txt
open_clip_torch
pymilvus
pydantic
